Product Page adding language change, redacting api-s
