+++ src/pages/ProductPage/ProductPage.module.scss
@@ -38,16 +38,3 @@
   margin-bottom: 32px;
   grid-column: 1/-1;
 }
-
-.productSlider {
-  grid-column: 1/-1;
-  margin-bottom: 56px;
-
-  @include mixines.tablet {
-    margin-bottom: 64px;
-  }
-
-  @include mixines.desctop {
-    margin-bottom: 80px;
-  }
-}
